change for asset

+++ jp/admin/assets.php
@@ -571,7 +571,7 @@ if(isset($_GET['pid'])&&$_GET['pid']!=''){
 
               $products_quantity_value = $info_value['products_quantity'] - ($products_quantity_num - $category_asset_arr[$key]['real_all_product']);
             }
-            $info_str_temp .= '<td align="right">'.$products_quantity_value.TEXT_ROW.'</td>';
+            $info_str_temp .= '<td align="right">'.(int)$products_quantity_value.TEXT_ROW.'</td>';
             $info_str_temp .= '<td align="right">'.($currencies->format(abs($info_value['final_price']))=='0'.TEXT_MONEY_SYMBOL?show_effective_number(abs($info_value['final_price'])):$currencies->format(abs($info_value['final_price']))).'</td>';
             $info_str_temp .= '<td align="right">'.($currencies->format(abs($products_quantity_value*$info_value['final_price']))=='0'.TEXT_MONEY_SYMBOL?show_effective_number(abs($products_quantity_value*$info_value['final_price'])):$currencies->format(abs($products_quantity_value*$info_value['final_price']))).'</td>';
             $info_str_temp .= '</tr>';
@@ -595,7 +595,7 @@ if(isset($_GET['pid'])&&$_GET['pid']!=''){
           echo $category_asset_arr[$key]['categories_name'];
           echo "</td>";
           echo "<td align='right'>";
-          echo $category_asset_arr[$key]['quantity_all_product'].TEXT_ROW;
+          echo (int)$category_asset_arr[$key]['quantity_all_product'].TEXT_ROW;
           echo "</td>";
           echo "<td align='right'>";
           echo $currencies->format($products_price_total/$products_quantity_total)=='0'.TEXT_MONEY_SYMBOL?show_effective_number($products_price_total/$products_quantity_total):$currencies->format($products_price_total/$products_quantity_total);
@@ -628,7 +628,7 @@ if(isset($_GET['pid'])&&$_GET['pid']!=''){
             echo tep_get_products_name($p_value);
             echo "</td>";
             echo "<td align='right'>";
-            echo tep_get_quantity($p_value).TEXT_ROW;
+            echo (int)tep_get_quantity($p_value).TEXT_ROW;
             echo "</td>";
             echo "<td align='right'>";
             echo $currencies->format($p_temp_price_total/$p_temp_quantity)=='0'.TEXT_MONEY_SYMBOL?show_effective_number($p_temp_price_total/$p_temp_quantity):$currencies->format($p_temp_price_total/$p_temp_quantity);
@@ -669,7 +669,7 @@ if(isset($_GET['pid'])&&$_GET['pid']!=''){
 
               $products_quantity_value = $info_value['products_quantity'] - ($products_quantity_num - $all_product[$k]['quantity_all_product']);
             }
-            $products_info_str .= '<td align="right">'.$products_quantity_value.TEXT_ROW.'</td>';
+            $products_info_str .= '<td align="right">'.(int)$products_quantity_value.TEXT_ROW.'</td>';
             $products_info_str .= '<td
               align="right">'.($currencies->format(abs($info_value['final_price']))=='0'.TEXT_MONEY_SYMBOL?show_effective_number(abs($info_value['final_price'])):$currencies->format(abs($info_value['final_price']))).'</td>';
             $products_info_str .= '<td align="right">'.($currencies->format(abs($products_quantity_value*$info_value['final_price']))=='0'.TEXT_MONEY_SYMBOL?show_effective_number(abs($products_quantity_value*$info_value['final_price'])):$currencies->format(abs($products_quantity_value*$info_value['final_price']))).'</td>';
@@ -695,7 +695,7 @@ if(isset($_GET['pid'])&&$_GET['pid']!=''){
         echo $all_product[$k]['products_name'];
         echo "</td>";
         echo "<td align='right'>";
-        echo $all_product[$k]['products_real_quantity'].TEXT_ROW;
+        echo (int)$all_product[$k]['products_real_quantity'].TEXT_ROW;
         echo "</td>";
         echo "<td align='right'>";
         echo $currencies->format($products_price_total/$products_quantity_total)=='0'.TEXT_MONEY_SYMBOL?show_effective_number($products_price_total/$products_quantity_total):$currencies->format($products_price_total/$products_quantity_total);
@@ -730,7 +730,7 @@ if(isset($_GET['pid'])&&$_GET['pid']!=''){
       echo TEXT_SUM_PRODUCT;
       echo "</td>";
       echo "<td id='info_value_td' class='assets_bottom_info'>";
-      echo $all_quantity.TEXT_ROW;
+      echo (int)$all_quantity.TEXT_ROW;
       echo "</td>";
       echo "</tr>";
       echo "<tr class='assets_c'>";
@@ -758,7 +758,7 @@ if(isset($_GET['pid'])&&$_GET['pid']!=''){
       echo TEXT_SUM_PRODUCT;
       echo "</td>";
       echo "<td class='asstes_easy_text'>";
-      echo $all_quantity.TEXT_ROW;
+      echo (int)$all_quantity.TEXT_ROW;
       echo "</td>";
       echo "</tr>";
       echo "<tr class='assets_c'>";
